Attach an existing AKS cluster to a workspace

If you already have an AKS cluster in your Azure subscription, and it is version 1.12.##, you can attach it to your workspace to use for deployments. The existing AKS cluster can be in a different Azure region than your workspace.

If you want to secure your AKS cluster using an Azure Virtual Network, you must create the virtual network first. For more information, see Secure Azure ML experimentation and inference jobs within an Azure Virtual Network

attach_aks_compute(workspace, cluster_name, resource_id = NULL,
  resource_group = NULL)

Arguments

workspace
The Workspace object to attach the AKS cluster to.

cluster_name
A string of the name for the cluster.

resource_id
A string of the resource ID for the AKS cluster being attached.

resource_group
A string of the resource group in which the AKS cluster is located.

Value

The AksCompute object.

Examples

ws <- load_workspace_from_config()
compute_target <- attach_aks_compute(ws,
                                     cluster_name = 'mycluster',
                                     resource_id = 'myresourceid',
                                     resource_group = 'myresourcegroup')