Create AmlCompute

Create AmlCompute

create_aml_compute(workspace, cluster_name, vm_size,
  vm_priority = "dedicated", min_nodes = 0, max_nodes = NULL,
  idle_seconds_before_scaledown = NULL, admin_username = NULL,
  admin_user_password = NULL, admin_user_ssh_key = NULL,
  vnet_resourcegroup_name = NULL, vnet_name = NULL,
  subnet_name = NULL, tags = NULL, description = NULL)

Arguments

workspace
workspace object

cluster_name
cluster name

vm_size
Size of agent VMs. More details can be found here: https://aka.ms/azureml-vm-details. Note that not all sizes are available in all regions, as detailed in the previous link.

vm_priority
dedicated or lowpriority VMs. If not specified, will default to dedicated.

min_nodes
Minimum number of nodes to use on the cluster. If not specified, will default to 0

max_nodes
Maximum number of nodes to use on the cluster

idle_seconds_before_scaledown
Node idle time in seconds before scaling down the cluster

admin_username
Name of the administrator user account which can be used to SSH into nodes

admin_user_password
Password of the administrator user account

admin_user_ssh_key
SSH public key of the administrator user account

vnet_resourcegroup_name
Name of the resource group where the virtual network is located

vnet_name
Name of the virtual network

subnet_name
Name of the subnet inside the vnet

tags
A dictionary of key value tags to provide to the compute object

description
A description to provide to the compute object