Deploy a Webservice from zero or more model objects.

Deploy a Webservice from zero or more model objects.

deploy_model(workspace, name, models, inference_config,
  deployment_config = NULL, deployment_target = NULL)

Arguments

workspace
A Workspace object to associate the Webservice with

name
The name to give the deployed service. Must be unique to the workspace, only consist of lowercase letters, numbers, or dashes, start with a letter, and be between 3 and 32 characters long.

models
A list of model objects. Can be an empty list.

inference_config
An InferenceConfig object used to determine required model properties.

deployment_config
A WebserviceDeploymentConfiguration used to configure the webservice. If one is not provided, an empty configuration object will be used based on the desired target.

deployment_target
A azureml.core.ComputeTarget to deploy the Webservice to. As Azure Container Instances has no associated azureml.core.ComputeTarget, leave this parameter as None to deploy to Azure Container Instances.