Update the Webservice with provided properties. Values left as None will remain unchanged in this Webservice.

Update the Webservice with provided properties. Values left as None will remain unchanged in this Webservice.

update_aks_webservice(webservice, autoscale_enabled = NULL,
  autoscale_min_replicas = NULL, autoscale_max_replicas = NULL,
  autoscale_refresh_seconds = NULL,
  autoscale_target_utilization = NULL, auth_enabled = NULL,
  cpu_cores = NULL, memory_gb = NULL, enable_app_insights = NULL,
  scoring_timeout_ms = NULL, replica_max_concurrent_requests = NULL,
  max_request_wait_time = NULL, num_replicas = NULL, tags = NULL,
  properties = NULL, description = NULL, models = NULL,
  inference_config = NULL, gpu_cores = NULL, period_seconds = NULL,
  initial_delay_seconds = NULL, timeout_seconds = NULL,
  success_threshold = NULL, failure_threshold = NULL,
  namespace = NULL, token_auth_enabled = NULL)

Arguments

webservice
AciWebservice object.

autoscale_enabled
Enable or disable autoscaling of this Webservice

autoscale_min_replicas
The minimum number of containers to use when autoscaling this Webservice

autoscale_max_replicas
The maximum number of containers to use when autoscaling this Webservice

autoscale_refresh_seconds
How often the autoscaler should attempt to scale this Webservice

autoscale_target_utilization
The target utilization (in percent out of 100) the autoscaler should attempt to maintain for this Webservice

auth_enabled
Whether or not to enable auth for this Webservice

cpu_cores
The number of cpu cores to allocate for this Webservice. Can be a decimal

memory_gb
The amount of memory (in GB) to allocate for this Webservice. Can be a decimal

enable_app_insights
Whether or not to enable Application Insights logging for this Webservice

scoring_timeout_ms
A timeout to enforce for scoring calls to this Webservice

replica_max_concurrent_requests
The number of maximum concurrent requests per node to allow for this Webservice

max_request_wait_time
The maximum amount of time a request will stay in the queue (in milliseconds) before returning a 503 error

num_replicas
The number of containers to allocate for this Webservice

tags
Dictionary of key value tags to give this Webservice. Will replace existing tags.

properties
Dictionary of key value properties to add to existing properties dictionary

description
A description to give this Webservice

models
A list of Model objects to package with the updated service

inference_config
An InferenceConfig object used to provide the required model deployment properties.

gpu_cores
The number of gpu cores to allocate for this Webservice

period_seconds
How often (in seconds) to perform the liveness probe. Default to 10 seconds. Minimum value is 1.

initial_delay_seconds
Number of seconds after the container has started before liveness probes are initiated.

timeout_seconds
Number of seconds after which the liveness probe times out. Defaults to 1 second. Minimum value is 1.

success_threshold
Minimum consecutive successes for the liveness probe to be considered successful after having failed. Defaults to 1. Minimum value is 1.

failure_threshold
When a Pod starts and the liveness probe fails, Kubernetes will try failureThreshold times before giving up. Defaults to 3. Minimum value is 1.

namespace
The Kubernetes namespace in which to deploy this Webservice: up to 63 lowercase alphanumeric ('a'-'z', '0'-'9') and hyphen ('-') characters. The first and last characters cannot be hyphens.

token_auth_enabled
Whether or not to enable Token auth for this Webservice. If this is enabled, users can access this Webservice by fetching access token using their Azure Active Directory credentials. Defaults to FALSE